feat: web metro resolver
