+++ packages/uniwind/package.json
@@ -29,13 +29,9 @@
             "types": "./dist/module/index.d.ts",
             "default": "./dist/common/index.js"
         },
-        "./components": {
-            "react-native": "./src/components/index.native.ts",
-            "import": "./dist/module/components/index.js",
-            "default": "./dist/common/components/index.js"
-        },
+        "./components": "./src/components/index.ts",
         "./components/*": {
-            "react-native": "./dist/common/components/native/*.js",
+            "react-native": "./src/components/native/*.tsx",
             "import": "./dist/module/components/web/*.js",
             "default": "./dist/common/components/web/*.js"
         },
