+++ packages/uniwind/src/components/web/ActivityIndicator.tsx
@@ -14,3 +14,5 @@ export const ActivityIndicator = copyComponentProperties(RNActivityIndicator, (p
         />
     )
 })
+
+export default ActivityIndicator
