+++ packages/uniwind/src/components/web/Image.tsx
@@ -14,3 +14,5 @@ export const Image = copyComponentProperties(RNImage, (props: ImageProps) => {
         />
     )
 })
+
+export default Image
