+++ packages/uniwind/src/components/web/ImageBackground.tsx
@@ -15,3 +15,5 @@ export const ImageBackground = copyComponentProperties(RNImageBackground, (props
         />
     )
 })
+
+export default ImageBackground
