+++ packages/uniwind/src/components/web/InputAccessoryView.ts
@@ -1,0 +1,2 @@
+import './rnw'
+export { InputAccessoryView } from 'react-native'
