+++ packages/uniwind/src/components/web/Modal.tsx
@@ -10,3 +10,5 @@ export const Modal = copyComponentProperties(RNModal, (props: ModalProps) => {
         />
     )
 })
+
+export default Modal
