+++ packages/uniwind/src/components/web/Pressable.tsx
@@ -10,3 +10,5 @@ export const Pressable = copyComponentProperties(RNPressable, (props: PressableP
         />
     )
 })
+
+export default Pressable
