+++ packages/uniwind/src/components/web/RefreshControl.tsx
@@ -10,3 +10,5 @@ export const RefreshControl = copyComponentProperties(RNRefreshControl, (props: 
         />
     )
 })
+
+export default RefreshControl
