+++ packages/uniwind/src/components/web/SafeAreaView.tsx
@@ -10,3 +10,5 @@ export const SafeAreaView = copyComponentProperties(RNSafeAreaView, (props: View
         />
     )
 })
+
+export default SafeAreaView
