+++ packages/uniwind/src/components/web/ScrollView.tsx
@@ -11,3 +11,5 @@ export const ScrollView = copyComponentProperties(RNScrollView, (props: ScrollVi
         />
     )
 })
+
+export default ScrollView
