+++ packages/uniwind/src/components/web/Text.tsx
@@ -10,3 +10,5 @@ export const Text = copyComponentProperties(RNText, (props: TextProps) => {
         />
     )
 })
+
+export default Text
