+++ packages/uniwind/src/components/web/TextInput.tsx
@@ -10,3 +10,5 @@ export const TextInput = copyComponentProperties(RNTextInput, (props: TextInputP
         />
     )
 })
+
+export default TextInput
