+++ packages/uniwind/src/components/web/TouchableNativeFeedback.tsx
@@ -1,0 +1,2 @@
+import './rnw'
+export { TouchableNativeFeedback } from 'react-native'
