+++ packages/uniwind/src/components/web/TouchableWithoutFeedback.tsx
@@ -10,3 +10,5 @@ export const TouchableWithoutFeedback = copyComponentProperties(RNTouchableWitho
         />
     )
 })
+
+export default TouchableWithoutFeedback
