+++ packages/uniwind/src/components/web/View.tsx
@@ -10,3 +10,5 @@ export const View = copyComponentProperties(RNView, (props: ViewProps) => {
         />
     )
 })
+
+export default View
